refactor: vercel ai sdk for custom-testcase
