+++ apps/nova/src/app/[locale]/(dashboard)/(admin)/problems/[problemId]/prompt-form.tsx
@@ -78,7 +78,7 @@ export default function PromptForm({ problem }: Props) {
         {
           method: 'POST',
           headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ prompt }),
+          body: JSON.stringify({ prompt: prompt, id: problem.id }),
         }
       );
